Move implementation of Client
